Added 3 tabs on admin accounts page
